requirements.txt
langchain
langchain-community
owlready2
pandas
rdflib
langchain-openai
python-dotenv
google-generativeai
google-cloud-aiplatform
rdflib